do not pass static functions as parameter of exporter
