+++ packages/ra-ui-materialui/src/index.js
@@ -6,5 +6,6 @@ export * from './field';
 export * from './input';
 export * from './layout';
 export * from './list';
+export * from './util';
 export Link from './Link';
 export defaultTheme from './defaultTheme';
