+++ packages/ra-ui-materialui/src/util/index.js
@@ -1,0 +1,1 @@
+export downloadCSV from './downloadCSV';
